Move status panel separately.
