+++ assets/src/edit-story/components/inspector/document/documentInspector.js
@@ -15,160 +15,14 @@
  */
 
 /**
- * External dependencies
- */
-import { useCallback, useEffect } from 'react';
-import styled from 'styled-components';
-
-/**
- * WordPress dependencies
- */
-import { __ } from '@wordpress/i18n';
-
-/**
  * Internal dependencies
  */
-import { useStory } from '../../../app/story';
-import { SimplePanel } from '../../panels/panel';
-import { Button, TextInput, Row, DropDown, HelperText } from '../../form';
-import useInspector from '../useInspector';
-import { PublishPanel, SlugPanel } from '../../panels';
-
-const BoxedTextInput = styled(TextInput)`
-  padding: 6px 6px;
-  border-radius: 4px;
-  flex-grow: 1;
-  &:focus {
-    background-color: #fff;
-  }
-`;
+import { PublishPanel, SlugPanel, StatusPanel } from '../../panels';
 
 function DocumentInspector() {
-  const {
-    actions: { loadStatuses, loadUsers },
-    state: { statuses },
-  } = useInspector();
-
-  const {
-    state: {
-      meta: { isSaving },
-      story: { status, password },
-      capabilities,
-    },
-    actions: { updateStory, deleteStory },
-  } = useStory();
-
-  useEffect(() => {
-    loadStatuses();
-    loadUsers();
-  });
-
-  const passwordProtected = 'protected';
-  const visibilityOptions = statuses.filter(({ value }) =>
-    ['draft', 'publish', 'private'].includes(value)
-  );
-
-  // @todo Add this back once we have FE implementation, too.
-  /*visibilityOptions.push({
-    name: __('Password Protected', 'web-stories'),
-    value: passwordProtected,
-  });*/
-
-  // @todo this should still allow showing the moment where the warning is red, currently just doesn't allow adding more.
-  const handleChangePassword = useCallback(
-    (value) => {
-      if (value.length <= 20) {
-        updateStory({
-          properties: { password: value },
-        });
-      }
-    },
-    [updateStory]
-  );
-
-  const isValidPassword = (value) => {
-    return value && value.length <= 20;
-  };
-
-  // @todo Design details.
-  const handleChangeVisibility = useCallback(
-    (value) => {
-      // If password protected but no password, do nothing.
-      if (value === passwordProtected && !isValidPassword(value)) {
-        return;
-      }
-      // If password protected, keep the previous status.
-      const properties =
-        passwordProtected === status
-          ? { password }
-          : {
-              status: value,
-              password: '',
-            };
-      updateStory({ properties });
-    },
-    [password, status, updateStory]
-  );
-
-  const handleRemoveStory = useCallback(
-    (evt) => {
-      deleteStory();
-      evt.preventDefault();
-    },
-    [deleteStory]
-  );
-
-  const getStatusValue = (value) => {
-    // Always display protected visibility, independent of the status.
-    if (password && password.length) {
-      return passwordProtected;
-    }
-    // Display as public even if scheduled for future, private post can't be scheduled.
-    if ('future' === value) {
-      return 'publish';
-    }
-    return value;
-  };
-
   return (
     <>
-      <SimplePanel
-        name="status"
-        title={__('Status & Visibility', 'web-stories')}
-      >
-        {capabilities && capabilities.hasPublishAction && statuses && (
-          <>
-            <Row>
-              <DropDown
-                ariaLabel={__('Visibility', 'web-stories')}
-                options={visibilityOptions}
-                disabled={isSaving}
-                value={getStatusValue(status)}
-                onChange={handleChangeVisibility}
-                isDocumentPanel={true}
-              />
-            </Row>
-            {passwordProtected === status && (
-              <>
-                <Row>
-                  <BoxedTextInput
-                    label={__('Password', 'web-stories')}
-                    value={password}
-                    onChange={handleChangePassword}
-                    placeholder={__('Enter a password', 'web-stories')}
-                  />
-                </Row>
-                <HelperText isWarning={password && password.length > 20}>
-                  {__('Must not exceed 20 characters', 'web-stories')}
-                </HelperText>
-              </>
-            )}
-          </>
-        )}
-        <Button onClick={handleRemoveStory} fullWidth>
-          {__('Move to trash', 'web-stories')}
-        </Button>
-      </SimplePanel>
+      <StatusPanel />
       <PublishPanel />
       <SlugPanel />
     </>
